third_party/blink/web_tests/wpt_internal/app-history/navigate-event/navigate-location.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<form id="form" action=""></form>
<script>
async_test(t => {
  appHistory.onnavigate = t.step_func_done(e => {
    assert_true(e.cancelable);
    assert_true(e.canRespond);
    assert_false(e.userInitiated);
    assert_true(e.hashChange);
    assert_equals(new URL(e.destination.url).hash, "#1");
    assert_true(e.destination.sameDocument);
    assert_equals(e.formData, null);
  });
  location.href = "#1";
}, "location API fires navigate event");
</script>
